be_flask/requirements.txt
arch==6.1.0
architect==0.5.6
click==7.1.1
cycler==0.10.0
Cython==0.29.16
Flask==1.0.2
Flask-Cors==3.0.8
Flask-WTF==0.14.2
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.5.0
numpy==1.22.3  # Adjusted for better compatibility
opencv-python==4.5.5.64  # Updated to a compatible version
pandas==1.4.2  # Updated to a compatible version
patsy==0.5.2
Pillow==8.4.0  # Updated to a compatible version
property-cached==1.6.4
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2020.1
scipy==1.8.0  # Updated to a compatible version
six==1.14.0
statsmodels==0.13.2  # Reverted back to a stable version
torch==1.12.0
torchvision==0.11.3  # Updated to a compatible version
Werkzeug==0.15.5
WTForms==2.3.0
